refactor: modularize storage adapter into organized submodules
Split monolithic storage_adapter.rs (3203 lines) into clean module structure:
- traits.rs: EvmStateStore trait definition
- types/: Domain types (account_aext, freeze, vote, witness, state_change)
- inmem.rs: In-memory implementation for testing
- engine_backed.rs: Production StorageEngine-backed implementation
- resource_tracker.rs: Tron power and freeze ledger tracking
- utils.rs: Shared utilities (varint encoding, protobuf helpers)

This improves maintainability, reduces compilation time, and provides clearer
separation of concerns for EVM state management in the Rust execution backend.
